// EvaluatorService/src/interfaces/evaluation.interface.ts
export interface TestCase {
    _id: string;
    input: string;
    output: string;
}

export interface Problem {
    id: string;
    title: string;
    description: string;
    difficulty: string;
    editorial: string;
    testcases: TestCase[];
    createdAt: Date;
    updatedAt: Date;
}

export interface EvaluationJob {
    submissionId: string;
    code: string;
    language: 'python' | 'cpp';
    problem: Problem;
}

export interface EvaluationResult {
    status: string;
    output: string | undefined;
}